Ui And User manage

+++ app/views/Admin/categoryList.php
@@ -28,10 +28,14 @@ $i = 0;
 		<td><?php echo $i; ?></td>
 		<td><?php echo $value['cat_name']; ?></td>
 		<td><?php echo $value['title']; ?></td>
+  <?php if(Session::get('level') == 1){ ?>
 		<td>
 			<a href="<?php echo BASE_URL ?>/Admin/editCat/<?php echo $value['id']; ?>">Edit</a> ||
 			<a onclick="return confirm('Are Your Sure to Delete Data!!')" href="<?php echo BASE_URL ?>/Admin/deleteCategoryById/<?php echo $value['id']; ?>">Delete</a>
 		</td>
+ <?php }else{ ?>
+     <td>Not Parmited</td>
+  <?php } ?>
 	</tr>
 <?php } ?>
 </table>
